project/project-16-AI-Based Virtual Assistant/requirements.txt
sounddevice
numpy
scipy
google-generativeai
streamlit
huggingface_hub
